+++ components/src/main/java/com/mercadolibre/android/andesui/textfield/AndesTextfield.kt
@@ -7,6 +7,7 @@ import android.text.Editable
 import android.text.InputFilter
 import android.text.InputType
 import android.text.TextWatcher
+import android.text.method.DigitsKeyListener
 import android.util.AttributeSet
 import android.util.TypedValue
 import android.view.LayoutInflater
@@ -17,6 +18,7 @@ import android.widget.TextView
 import com.facebook.drawee.view.SimpleDraweeView
 import com.mercadolibre.android.andesui.R
 import com.mercadolibre.android.andesui.button.AndesButton
+import com.mercadolibre.android.andesui.color.AndesColor
 import com.mercadolibre.android.andesui.color.toAndesColor
 import com.mercadolibre.android.andesui.icons.IconProvider
 import com.mercadolibre.android.andesui.textfield.content.AndesTextfieldLeftContent
@@ -25,6 +27,8 @@ import com.mercadolibre.android.andesui.textfield.factory.AndesTextfieldAttrs
 import com.mercadolibre.android.andesui.textfield.factory.AndesTextfieldAttrsParser
 import com.mercadolibre.android.andesui.textfield.factory.AndesTextfieldConfiguration
 import com.mercadolibre.android.andesui.textfield.factory.AndesTextfieldConfigurationFactory
+import com.mercadolibre.android.andesui.textfield.maskTextField.TextFieldMask
+import com.mercadolibre.android.andesui.textfield.maskTextField.TextFieldMaskWatcher
 import com.mercadolibre.android.andesui.textfield.state.AndesTextfieldState
 import com.mercadolibre.android.andesui.textfield.state.AndesTextfieldState.*
 import com.mercadolibre.android.andesui.utils.buildColoredAndesBitmapDrawable
@@ -159,11 +163,12 @@ class AndesTextfield : ConstraintLayout {
     private lateinit var iconComponent: SimpleDraweeView
     private lateinit var leftComponent: FrameLayout
     private lateinit var rightComponent: FrameLayout
+    private var maskWatcher: TextFieldMaskWatcher? = null
 
     @Suppress("unused")
     constructor(context: Context) : super(context) {
         initAttrs(LABEL_DEFAULT, HELPER_DEFAULT, PLACEHOLDER_DEFAULT, COUNTER_DEFAULT,
-                  STATE_DEFAULT, LEFT_COMPONENT_DEFAULT, RIGHT_COMPONENT_DEFAULT, INPUT_TYPE_DEFAULT)
+            STATE_DEFAULT, LEFT_COMPONENT_DEFAULT, RIGHT_COMPONENT_DEFAULT, INPUT_TYPE_DEFAULT)
     }
 
     constructor(
@@ -206,7 +211,7 @@ class AndesTextfield : ConstraintLayout {
     ) {
         val showCounter = SHOW_COUNTER_DEFAULT
         andesTextfieldAttrs = AndesTextfieldAttrs(
-                label, helper, placeholder, counter, showCounter, state, leftContent, rightContent, inputType
+            label, helper, placeholder, counter, showCounter, state, leftContent, rightContent, inputType
         )
         val config = AndesTextfieldConfigurationFactory.create(context, andesTextfieldAttrs)
         setupComponents(config)
@@ -371,9 +376,24 @@ class AndesTextfield : ConstraintLayout {
             }
 
             override fun onTextChanged(charSequence: CharSequence?, start: Int, before: Int, count: Int) {
-                counterComponent.text = resources.getString(R.string.andes_textfield_counter_text, charSequence!!.length, config.counterLength)
+
+                val textWithoutMask = cleanText(charSequence.toString())
+
+                counterComponent.text = resources.getString(R.string.andes_textfield_counter_text,
+                    textWithoutMask?.length, config.counterLength)
             }
         })
+    }
+
+    /**
+     * if use maskWatcher will remove masks chars
+     */
+    private fun cleanText(text: String): String? {
+        return if (maskWatcher == null) {
+            text
+        } else {
+            maskWatcher?.cleanMask(text)
+        }
     }
 
     /**
@@ -484,14 +504,24 @@ class AndesTextfield : ConstraintLayout {
     /**
      * Set the right content to icon and provides an interface to give the icon path.
      */
-    fun setRightIcon(iconPath: String) {
+    fun setRightIcon(iconPath: String, listener: OnClickListener? = null, colorIcon: Int? = R.color.andes_gray_800) {
         rightContent = AndesTextfieldRightContent.ICON
         val rightIcon: SimpleDraweeView = rightComponent.getChildAt(0) as SimpleDraweeView
+
+        var color: AndesColor? = null
+        if (colorIcon != null) {
+            color = colorIcon.toAndesColor()
+        }
+
         rightIcon.setImageDrawable(buildColoredAndesBitmapDrawable(
-                IconProvider(context).loadIcon(iconPath) as BitmapDrawable,
-                context,
-                color = R.color.andes_gray_800.toAndesColor())
+            IconProvider(context).loadIcon(iconPath) as BitmapDrawable,
+            context,
+            color = color)
         )
+
+        if (listener != null) {
+            rightIcon.setOnClickListener(listener)
+        }
     }
 
     /**
@@ -501,9 +531,9 @@ class AndesTextfield : ConstraintLayout {
         leftContent = AndesTextfieldLeftContent.ICON
         val leftIcon: SimpleDraweeView = leftComponent.getChildAt(0) as SimpleDraweeView
         leftIcon.setImageDrawable(buildColoredAndesBitmapDrawable(
-                IconProvider(context).loadIcon(iconPath) as BitmapDrawable,
-                context,
-                color = R.color.andes_gray_450.toAndesColor())
+            IconProvider(context).loadIcon(iconPath) as BitmapDrawable,
+            context,
+            color = R.color.andes_gray_450.toAndesColor())
         )
     }
 
@@ -523,6 +553,35 @@ class AndesTextfield : ConstraintLayout {
         rightContent = AndesTextfieldRightContent.SUFFIX
         val suffix: TextView = rightComponent.getChildAt(0) as TextView
         suffix.text = text
+    }
+
+    /**
+     * Set mask in text field.
+     * @param mask mask pattern ##-###-######-# digits sample 0123456789-
+     */
+    fun setTextFieldMask(mask: TextFieldMask, listener: TextFieldMaskWatcher.OnTextChange?) {
+
+        if (maskWatcher == null) {
+            maskWatcher = TextFieldMaskWatcher(mask.mask, listener)
+            textComponent.addTextChangedListener(maskWatcher)
+        }
+
+        maskWatcher!!.setMask(mask.mask)
+
+        counter = maskWatcher?.getMaxLength()!!
+        textComponent.filters = arrayOf<InputFilter>(InputFilter.LengthFilter((mask.mask.length)))
+
+        if (mask.digits != null && mask.digits.isNotEmpty()) {
+            textComponent.keyListener = DigitsKeyListener.getInstance(mask.digits)
+        }
+    }
+
+    /**
+     * remove mask in text field.
+     */
+    fun clearMask() {
+        textComponent.removeTextChangedListener(maskWatcher)
+        maskWatcher = null
     }
 
     fun requestFocusOnTextField() {