+++ components/src/main/java/com/mercadolibre/android/andesui/textfield/maskTextField/TextFieldMask.kt
@@ -1,0 +1,8 @@
+package com.mercadolibre.android.andesui.textfield.maskTextField
+
+import java.io.Serializable
+
+data class TextFieldMask(
+    val mask: String,
+    val digits: String? = null
+) : Serializable