+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v2.7.1
+## Added
+- Mask to textField
+- Listener click in right icon
+- Color change to right icon
+
 # v2.7.0
 ## Added
 - Link action in AndesMessage Component.